apps/netconnector: add lib class NetStubResponder
NetStubResponder is used in conjunction with a proxy and stub to manage
stubs on the service (stub) end. An object that wants to expose itself to
remote parties via netconnector creates a NetStubResponder. The responder
registers with netconnector as a responding service and instantiates a
new stub when a connection is requested.

Change-Id: Ief1c46bd9041dd18ac745fad68b67236984a926e