apps/netconnector: wrap ip addresses and ports
This CL introduces three types: IpPort, IpAddress and SocketAddress,
which represent an ip port, a v4 or v6 address and a v4 or v6 address
and port, respectively. These classes leverage c++ goodness to reduce
the probability of address-related coding errors sneaking past the
compiler. These types will be used extensively in the new mDNS
implementation.

Change-Id: I33b53348a577abcfe0588826e5c2ab0e38af5db9