apps/netconnector: DNS message changes
This CL makes various changes to the way DNS messages are represented
in memory. Some changes are about making message building easier. The
biggest change is that questions and records, the coin of the realm
in mDNS, are referenced using shared_ptr, so they can more easily be
passed around without copying or other awkwardness.

Change-Id: I161148204b3f577d1f4283d1cd1d5983c08fa5a6