apps/netconnector: mdns transceiver changes
This CL makes a few changes to MdnsTransceiver and subordinates:
- Startup failures are reported so we fail nicely with netstack2
- Address fixups are implemented
- v4 multicast destination addresses are converted to v6 as needed
- Misc other stuff
Address fixups are needed, because the executive-level components
of mDNS (yet to be checked in) don't know the actual addresses to
report when sending 'A' and 'AAAA' (v4 and v6 address) records.
These addresses are interface-specific and need to be filled in
on a per-interface basis.

Change-Id: I267f8c16a8e33e4272ec7aa29c15a2bd62daee46