apps/netconnector: use gethostname to get the host name
netconnector previously had its own idea of what the host name is
regardless of gethostname (which still is implemented properly).
There was a bunch of stuff in params, config and in the fidl
interface to set the host name.

This CL makes netconnector use gethostname. Because gethostname
always returns 'fuchsia' at the moment, there's some logic to
append part of the IP address for disambiguation purposes. This
will code will be removed once gethostname actually works.

Change-Id: I52573096e951bcbc3f94f5b5ad3b55f4b41a9e3e