apps/netconnector: implement resource renewal
Up to this point, the mDNS implementation has been pretending that
inbound TTLs were infinite. This means that when a service instance
is removed from the network without sending deletion records, the
instance subscriber would never know that the device had gone.

This CL implements resource renewal for resources we care about
long-term (just those cached by InstanceSubscriber for now). It
introduces a new agent, ResourceRenewer, that implements most of
this. ResourceRenewer keeps a list of resources that need to be
renewed. When the TTL of such a resource approaches, the renewer
begins to query for updated copies of that resource. When the
update arrives or the TTL expires, the renewer forgets about the
resource. In the expiration case, the renewer also sends that
resource to all other agents with a TTL of zero, indicating that
the resource should be removed whereever it's cached.

Agents like the InstanceSubscriber that need renewal services ask
for renewal whenever a new copy of a given resource arrives. With
this approach, there's no need for InstanceSubscriber to cancel
renewal requests, making the implementation more robust at the cost
of occaisionally attempting to renew a resource that's not really
interesting.

Change-Id: Iace03f2b759b087c0c5c173aee13245b16a561b3