apps/netconnector: get a unique host name during boot
This CL adds a temporary hack that makes netconnector poll its
GetHostName function until it returns something likely to be
unique on the LAN.

gethostname currently returns "fuchsia" in all cases. NetConnector
tries to dedup host names by asking the network stack for an IP
address and using one of the address bytes as a suffix.

When NetConnector is in the boot sequence and the NIC is a USB
device, the network stack doesn't enumerate the NIC in time for
this dedup to work.

This CL adds a hack that makes NetConnector poll its GetHostName
function until it returns something besides "fuchsia" or
"fuchsia-0", the latter occurring when the NIC has been
enumerated but DHCP hasn't yet issued an IP address.

Change-Id: Ie27bdf93718dd84f2090000a34bd8b313aba08a5