apps/netconnector: NET-75 handle service method arg errors
NetConnector was DCHECKing when a bad service name was submitted via
service methods such as PublishServiceInstance. This CL adds checks
in those methods and allows the requests to fail without crashing
the service.

The action taken when these errors are found is basically to ignore
the request. I'd rather close the channel in question, but I'm not
sure how to figure out which one it is. If there's a way to do this,
I'll submit a new CL.

Change-Id: I9e9e1bb7995bfe4eef75e4e3b976938407973636