apps/netconnector: make mDNS startup quieter
This CL makes NetConnector wait until a NIC is ready with an assigned
address before trying start to start mDNS. It also reduces the number
of messages that appear in the log as NetConnector waits for the
network to warm up.

The mDNS implementation should be responsible for handling the
arrival and departure of useful NICs, and should happily start before
the network is warmed up. At this point, though, we can't get a
reasonably unique host name before we have a host address, so we need
to wait for network warm-up before starting mDNS.

This CL checks whether the network is ready every 5 seconds until a
ready NIC is found. It generates no log messages while this is
happening. In addition to reducing log spam, this approach has the
advantage that mDNS will work properly even if it takes a long time
for a NIC to show up (e.g. user inserts a USB NIC well after boot)
and get an address (e.g. user connects network cable or configures
DHCP after boot).

This CL has been tested on NIC-less qemu and on real hardware with
delayed USB NIC insertion.

Change-Id: I856e29607fc0c93fdd6b6fd5aa6b98ed364bd618